// src/models/getAllUsersResponseModel.ts
export interface getAllUserRespModel {
    isAdmin: boolean
    isActive: boolean
    _id: string
    firstName: string
    lastName: string
    email: string
    dob?: string
    gender: string
    photoId?: string
    createdDate: string
    __v: number
    id: string
    phone?: string
    city?: string
    state?: string
    country?: string
    pincode?: string
    profession?: string
  }